backend/requirements.txt
# ESI-Online Backend - Dépendances de base
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
psycopg2-binary>=2.9,<3.0
jinja2>=3.1,<4.0
